finally removed: feature(generic_const_exprs)

+++ examples/b_2_time.rs
@@ -1,10 +1,7 @@
-#![allow(incomplete_features)]
-#![feature(generic_const_exprs)]
-
 use phases::{Array2d, System};
 
 // model parameters
-const N_ATOMS: u8 = 2;
+const N_ATOMS: usize = 2;
 type Atom = phases::NumAtom<N_ATOMS>;
 type Concentration = phases::NumC<N_ATOMS>;
 const WIDTH: usize = 200;
